% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{buildForward}
\alias{buildForward}
\title{Map from Indices to Label Pairs}
\usage{
buildForward(N)
}
\arguments{
\item{N}{Sample size, i.e., nrow(\eqn{\mathbf{X}})}
}
\value{
\eqn{N \times N} matrix whose entries record the index
corresponding  to the pair of labels (indexed by the matrix dims)
}
\description{
Builds a map from indexes to pairs of labels. This is
for caching distances, to avoid recomputing Hamming distances
especially when dealing with high-dimensional (large \eqn{P}) matrices.
}
\details{
Dependencies: None
}
