% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{buildReverse}
\alias{buildReverse}
\title{Map from Label Pairs to Indices}
\usage{
buildReverse(N)
}
\arguments{
\item{N}{Sample size, i.e., nrow(X)}
}
\value{
N x N matrix whose entries record the index
corresponding  to the pair of labels (indexed by the matrix dims)
}
\description{
Builds a map from pairs of labels to indexes. This is
for caching distances, to avoid recomputing Hamming distances
especially when dealing with high-dimensional (large P) matrices.
}
\details{
Dependencies: None
}
