\name{flintyR-package}
\alias{flintyR-package}
\alias{flintyR}
\docType{package}
\title{\packageTitle{flintyR}}
\description{\packageDescription{flintyR}}
\section{Package Content}{\packageIndices{flintyR}}
\author{\packageAuthor{flintyR}}
\section{Maintainer}{\packageMaintainer{flintyR}}
\keyword{package}
