% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{getChi2Weights}
\alias{getChi2Weights}
\title{Get Chi Square Weights}
\usage{
getChi2Weights(alpha, beta, gamma, N)
}
\arguments{
\item{alpha}{covariance matrix entry computed from getCov}

\item{beta}{covariance matrix entry computed from getCov}

\item{gamma}{covariance matrix entry computed from getCov}

\item{N}{The sample size, i.e., nrow(X) where X is the original dataset}
}
\value{
The weights \eqn{(w_1, w_2)}
}
\description{
Computes weights for the asymptotic random variable
from the \eqn{\alpha, \beta} and \eqn{\gamma} computed of data array \eqn{\mathbf{X}}.
}
\details{
This is used in the large \eqn{P} asymptotics of the permutation test.

Dependencies: None
}
