% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming_bitwise}
\alias{hamming_bitwise}
\title{Fast Bitwise Hamming Distance Vector Computation}
\usage{
hamming_bitwise(X)
}
\arguments{
\item{X}{binary matrix (IntegerMatrix class )}
}
\value{
vector of Hamming distances (NumericVector class)
}
\description{
Takes in a binary matrix \eqn{\mathbf{X}}, whose transpose \eqn{\mathbf{X}^T}
has \eqn{N} rows, and computes a vector recording all
\eqn{{N \choose 2}} pairwise Hamming distances of \eqn{\mathbf{X}^T},
ordered lexicographically.
}
\examples{
# t(X) = [[1,0], [0,1], [1,1]] --> output = [2,1,1]
}
