% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{indLargeP}
\alias{indLargeP}
\title{Approximate p-value for Test of Exchangeability (Assuming Large P)}
\usage{
indLargeP(X, p = 2)
}
\arguments{
\item{X}{The binary or real matrix on which to perform test of exchangeability}

\item{p}{The power p of \eqn{l_p^p}, i.e., \eqn{||x||_p^p = (x_1^p+...x_n^p)}}
}
\value{
The asymptotic p-value
}
\description{
Computes the large P asymptotic p-value for dataset X,
assuming its P features are independent.
}
\details{
This is the large P asymptotics of the permutation test.

Dependencies: getBinVStat, getRealVStat, getChi2Weights, weightedChi2P, getCov
}
