% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{getChi2Weights}
\alias{getChi2Weights}
\title{Get Chi Square Weights}
\usage{
getChi2Weights(alpha, beta, gamma, N)
}
\arguments{
\item{alpha}{covariance matrix entry computed from getCov}

\item{beta}{covariance matrix entry computed from getCov}

\item{gamma}{covariance matrix entry computed from getCov}

\item{N}{The sample size, i.e., nrow(X) where X is the original dataset}
}
\value{
The weights (w1, w2)
}
\description{
Computes weights for the asymptotic random variable
from the alpha, beta and gamma computed of data array X.
}
\details{
This is used in the large P asymptotics of the permutation test.

Dependencies: None
}
